Change messages transmitted between map and marker and their views
Now a Pin is managed by the MapView, and that's MapController which
creates a Marker, and then asks the MapView to create a Pin with this
new marker.

This indirectly allows to not rely on the fact that MapView is a
ScrollPane, and thus add control widgets *inside* the MapView by
changing its parent class without breaking Pin drawing.