Reorganize JavaFX objects in a more decoupled way
Now we have one global layout, which is instanciated in Main. MapView
gets it with the static getter, and adds itself into it.

MapView is a ScrollPane, so that its content can be scrolled. It
contains one StackPane, which is used in order to add more than one
child (such as the pins and the image). In this StackPane, we add the
image, and then the Pins when they are created.

You don't want to know how it was working before, this behavior is
waaay cleaner. Thrust me.