Use more consistently exceptions, even for signin/up
In fact, the interface of the various query controller is far more
consistent if we have not two methods where a boolean is returned to
indicate success, while the other return void.
Now, all methods return void and throw InvalidParameterException.

The benefit is more in the callers of these methods than in the
signature: we have try-catch everywhere, and no more boolean-if-else.